requirements-temp.txt
fastapi==0.105.0
uvicorn==0.24.0
tortoise-orm==0.20.0
asyncpg==0.29.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.4.2
email-validator==2.0.0
